refactor(flow): extract assemble-flow-share module + reuse shared esc/helpers
